fix module namespace of realm sync API
it isn't mounted in PVE::API2::AccessControl and it doesn't lives
anywhere in /access, so using that is just confusing.

Both, API and backend could simply move to manager, but as we already
got an api package here and it does somewhat fits into the topic lets
keep it here for now.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>