api: domains: fix setting only the realm authentication password
Previously, if one tried to only update e.g. the bind password for an
LDAP realm, it would fail with an rather unhelpful error message:

  # pveum realm modify ldap -password foo
  update auth server failed: no options specified

The root cause was that the `password` parameter was removed early from
the parameter object, which than would fail the check whether it is
empty or not.

Thus, additionally check if only `password` was specified and if so,
allow it.

Reported-by: Friedrich Weber <f.weber@proxmox.com>
Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>